add tests for XPath context using idValue
Test that XPath context size and position are set properly for relative
selectors. Do this by using the XPath `pos()` and `last()` functions
in an `idValue`, which is an XPath that evaluates to a string.

This relates to this discussion on the ITS IG list:
lists.w3.org/Archives/Public/public-i18n-its-ig/2013Jul/0037.html